Fix numFmt for Excel
Excel (at least up to 2016, possibly later) doesn't support using "NN"
for the day of the week.  "ddd" seems to work just as well.

Programs generally seem to espace spaces in number formats.  This
doesn't seem to change anything, but is at worst a harmless way to
behave more standardly.